requirements.txt
git+https://github.com/Rapptz/discord.py@bdea50e1db3c12a336c4395203f7694d21b176aa
pyyaml==5.1
